ariel_os_esp

Module peripherals

Types for the peripheral singletons.

Structs

ADC1unstable

ADC1 peripheral singleton

AESunstable

AES peripheral singleton

APB_SARADCunstable

APB_SARADC peripheral singleton

ASSIST_DEBUGunstable

ASSIST_DEBUG peripheral singleton

ATOMICunstable

ATOMIC peripheral singleton

BTunstable

BT peripheral singleton

DMAunstable

DMA peripheral singleton

DMA_CH0unstable

DMA_CH0 peripheral singleton

DMA_CH1unstable

DMA_CH1 peripheral singleton

DMA_CH2unstable

DMA_CH2 peripheral singleton

DSunstable

DS peripheral singleton

ECCunstable

ECC peripheral singleton

EFUSEunstable

EFUSE peripheral singleton

ETMunstable

ETM peripheral singleton

EXTMEMunstable

EXTMEM peripheral singleton

FLASHunstable

FLASH peripheral singleton

GPIOunstable

GPIO peripheral singleton

GPIO0

GPIO0 peripheral singleton

GPIO1

GPIO1 peripheral singleton

GPIO2

GPIO2 peripheral singleton

GPIO3

GPIO3 peripheral singleton

GPIO4

GPIO4 peripheral singleton

GPIO5

GPIO5 peripheral singleton

GPIO6

GPIO6 peripheral singleton

GPIO7

GPIO7 peripheral singleton

GPIO8

GPIO8 peripheral singleton

GPIO9

GPIO9 peripheral singleton

GPIO10

GPIO10 peripheral singleton

GPIO11

GPIO11 peripheral singleton

GPIO12

GPIO12 peripheral singleton

GPIO13

GPIO13 peripheral singleton

GPIO14

GPIO14 peripheral singleton

GPIO15

GPIO15 peripheral singleton

GPIO16

GPIO16 peripheral singleton

GPIO17

GPIO17 peripheral singleton

GPIO18

GPIO18 peripheral singleton

GPIO19

GPIO19 peripheral singleton

GPIO20

GPIO20 peripheral singleton

GPIO21

GPIO21 peripheral singleton

GPIO22

GPIO22 peripheral singleton

GPIO23

GPIO23 peripheral singleton

GPIO24

GPIO24 peripheral singleton

GPIO25

GPIO25 peripheral singleton

GPIO26

GPIO26 peripheral singleton

GPIO27

GPIO27 peripheral singleton

GPIO28

GPIO28 peripheral singleton

GPIO29

GPIO29 peripheral singleton

GPIO30

GPIO30 peripheral singleton

GPIO_SDunstable

GPIO_SD peripheral singleton

HINFunstable

HINF peripheral singleton

HMACunstable

HMAC peripheral singleton

HP_APMunstable

HP_APM peripheral singleton

HP_SYSunstable

HP_SYS peripheral singleton

I2C0

I2C0 peripheral singleton

I2C_ANA_MSTunstable

I2C_ANA_MST peripheral singleton

I2S0unstable

I2S0 peripheral singleton

IEEE802154unstable

IEEE802154 peripheral singleton

INTERRUPT_CORE0unstable

INTERRUPT_CORE0 peripheral singleton

INTPRIunstable

INTPRI peripheral singleton

IO_MUXunstable

IO_MUX peripheral singleton

LEDCunstable

LEDC peripheral singleton

LPWRunstable

LPWR peripheral singleton

LP_ANAunstable

LP_ANA peripheral singleton

LP_AONunstable

LP_AON peripheral singleton

LP_APMunstable

LP_APM peripheral singleton

LP_APM0unstable

LP_APM0 peripheral singleton

LP_CLKRSTunstable

LP_CLKRST peripheral singleton

LP_COREunstable

LP_CORE peripheral singleton

LP_I2C0unstable

LP_I2C0 peripheral singleton

LP_I2C_ANA_MSTunstable

LP_I2C_ANA_MST peripheral singleton

LP_IOunstable

LP_IO peripheral singleton

LP_PERIunstable

LP_PERI peripheral singleton

LP_TEEunstable

LP_TEE peripheral singleton

LP_TIMERunstable

LP_TIMER peripheral singleton

LP_UARTunstable

LP_UART peripheral singleton

LP_WDTunstable

LP_WDT peripheral singleton

MCPWM0unstable

MCPWM0 peripheral singleton

MEM2MEM1unstable

MEM2MEM1 peripheral singleton

MEM2MEM4unstable

MEM2MEM4 peripheral singleton

MEM2MEM5unstable

MEM2MEM5 peripheral singleton

MEM2MEM10unstable

MEM2MEM10 peripheral singleton

MEM2MEM11unstable

MEM2MEM11 peripheral singleton

MEM2MEM12unstable

MEM2MEM12 peripheral singleton

MEM2MEM13unstable

MEM2MEM13 peripheral singleton

MEM2MEM14unstable

MEM2MEM14 peripheral singleton

MEM2MEM15unstable

MEM2MEM15 peripheral singleton

MEM_MONITORunstable

MEM_MONITOR peripheral singleton

MODEM_LPCONunstable

MODEM_LPCON peripheral singleton

MODEM_SYSCONunstable

MODEM_SYSCON peripheral singleton

OTP_DEBUGunstable

OTP_DEBUG peripheral singleton

OptionalPeripherals

The Peripherals struct provides access to all of the hardware peripherals on the chip.

PARL_IOunstable

PARL_IO peripheral singleton

PAUunstable

PAU peripheral singleton

PCNTunstable

PCNT peripheral singleton

PCRunstable

PCR peripheral singleton

PLIC_MXunstable

PLIC_MX peripheral singleton

PMUunstable

PMU peripheral singleton

Peripherals

The Peripherals struct provides access to all of the hardware peripherals on the chip.

RMTunstable

RMT peripheral singleton

RNGunstable

RNG peripheral singleton

RSAunstable

RSA peripheral singleton

SHAunstable

SHA peripheral singleton

SLCHOSTunstable

SLCHOST peripheral singleton

SPI0unstable

SPI0 peripheral singleton

SPI1unstable

SPI1 peripheral singleton

SPI2

SPI2 peripheral singleton

SW_INTERRUPTunstable

SW_INTERRUPT peripheral singleton

SYSTEMunstable

SYSTEM peripheral singleton

SYSTIMERunstable

SYSTIMER peripheral singleton

TEEunstable

TEE peripheral singleton

TIMG0unstable

TIMG0 peripheral singleton

TIMG1unstable

TIMG1 peripheral singleton

TRACE0unstable

TRACE0 peripheral singleton

TSENSunstable

TSENS peripheral singleton

TWAI0unstable

TWAI0 peripheral singleton

TWAI1unstable

TWAI1 peripheral singleton

UART0

UART0 peripheral singleton

UART1

UART1 peripheral singleton

UHCI0unstable

UHCI0 peripheral singleton

USB_DEVICEunstable

USB_DEVICE peripheral singleton

WIFIunstable

WIFI peripheral singleton

Enums

Interrupt

Enumeration of all the interrupts.