ariel_os_esp::peripherals

Struct LP_ANA

#[non_exhaustive]
pub struct LP_ANA<'a> { /* private fields */ }

Available on crate features unstable only.

LP_ANA peripheral singleton

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

Implementations

impl LP_ANA<'_>

pub unsafe fn steal() -> LP_ANA<'_>

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub unsafe fn clone_unchecked(&self) -> LP_ANA<'_>

Unsafely clone this peripheral reference.

Safety

You must ensure that you’re only using one instance of this type at a time.

pub fn reborrow(&mut self) -> LP_ANA<'_>

Creates a new peripheral reference with a shorter lifetime.

Use this method if you would like to keep working with the peripheral after you dropped the driver that consumes this.

impl LP_ANA<'_>

pub const PTR: *const <Periph<RegisterBlock, 1611344896> as Deref>::Target = pac::LP_ANA::PTR

Pointer to the register block

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub const fn ptr() -> *const <Periph<RegisterBlock, 1611344896> as Deref>::Target

Return the pointer to the register block

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub const fn regs<'a>() -> &'a <Periph<RegisterBlock, 1611344896> as Deref>::Target

Return a reference to the register block

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

pub fn register_block( &self, ) -> &<Periph<RegisterBlock, 1611344896> as Deref>::Target

Return a reference to the register block

Stability

This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.