ariel_os_esp::peripherals

Struct OptionalPeripherals

pub struct OptionalPeripherals {

    pub GPIO0: Option<GPIO0<'static>>,
    pub GPIO1: Option<GPIO1<'static>>,
    pub GPIO2: Option<GPIO2<'static>>,
    pub GPIO3: Option<GPIO3<'static>>,
    pub GPIO4: Option<GPIO4<'static>>,
    pub GPIO5: Option<GPIO5<'static>>,
    pub GPIO6: Option<GPIO6<'static>>,
    pub GPIO7: Option<GPIO7<'static>>,
    pub GPIO8: Option<GPIO8<'static>>,
    pub GPIO9: Option<GPIO9<'static>>,
    pub GPIO10: Option<GPIO10<'static>>,
    pub GPIO11: Option<GPIO11<'static>>,
    pub GPIO12: Option<GPIO12<'static>>,
    pub GPIO13: Option<GPIO13<'static>>,
    pub GPIO14: Option<GPIO14<'static>>,
    pub GPIO15: Option<GPIO15<'static>>,
    pub GPIO16: Option<GPIO16<'static>>,
    pub GPIO17: Option<GPIO17<'static>>,
    pub GPIO18: Option<GPIO18<'static>>,
    pub GPIO19: Option<GPIO19<'static>>,
    pub GPIO20: Option<GPIO20<'static>>,
    pub GPIO21: Option<GPIO21<'static>>,
    pub GPIO22: Option<GPIO22<'static>>,
    pub GPIO23: Option<GPIO23<'static>>,
    pub GPIO27: Option<GPIO27<'static>>,
    pub AES: Option<AES<'static>>,
    pub APB_SARADC: Option<APB_SARADC<'static>>,
    pub ASSIST_DEBUG: Option<ASSIST_DEBUG<'static>>,
    pub ATOMIC: Option<ATOMIC<'static>>,
    pub DMA: Option<DMA<'static>>,
    pub DS: Option<DS<'static>>,
    pub ECC: Option<ECC<'static>>,
    pub EFUSE: Option<EFUSE<'static>>,
    pub EXTMEM: Option<EXTMEM<'static>>,
    pub GPIO: Option<GPIO<'static>>,
    pub GPIO_SD: Option<GPIO_SD<'static>>,
    pub HINF: Option<HINF<'static>>,
    pub HMAC: Option<HMAC<'static>>,
    pub HP_APM: Option<HP_APM<'static>>,
    pub HP_SYS: Option<HP_SYS<'static>>,
    pub I2C_ANA_MST: Option<I2C_ANA_MST<'static>>,
    pub I2C0: Option<I2C0<'static>>,
    pub I2S0: Option<I2S0<'static>>,
    pub IEEE802154: Option<IEEE802154<'static>>,
    pub INTERRUPT_CORE0: Option<INTERRUPT_CORE0<'static>>,
    pub INTPRI: Option<INTPRI<'static>>,
    pub IO_MUX: Option<IO_MUX<'static>>,
    pub LEDC: Option<LEDC<'static>>,
    pub LP_ANA: Option<LP_ANA<'static>>,
    pub LP_AON: Option<LP_AON<'static>>,
    pub LP_APM: Option<LP_APM<'static>>,
    pub LP_APM0: Option<LP_APM0<'static>>,
    pub LP_CLKRST: Option<LP_CLKRST<'static>>,
    pub LP_I2C0: Option<LP_I2C0<'static>>,
    pub LP_I2C_ANA_MST: Option<LP_I2C_ANA_MST<'static>>,
    pub LP_IO: Option<LP_IO<'static>>,
    pub LP_PERI: Option<LP_PERI<'static>>,
    pub LP_TEE: Option<LP_TEE<'static>>,
    pub LP_TIMER: Option<LP_TIMER<'static>>,
    pub LP_UART: Option<LP_UART<'static>>,
    pub LP_WDT: Option<LP_WDT<'static>>,
    pub LPWR: Option<LPWR<'static>>,
    pub MCPWM0: Option<MCPWM0<'static>>,
    pub MEM_MONITOR: Option<MEM_MONITOR<'static>>,
    pub MODEM_LPCON: Option<MODEM_LPCON<'static>>,
    pub MODEM_SYSCON: Option<MODEM_SYSCON<'static>>,
    pub OTP_DEBUG: Option<OTP_DEBUG<'static>>,
    pub PARL_IO: Option<PARL_IO<'static>>,
    pub PAU: Option<PAU<'static>>,
    pub PCNT: Option<PCNT<'static>>,
    pub PCR: Option<PCR<'static>>,
    pub PLIC_MX: Option<PLIC_MX<'static>>,
    pub PMU: Option<PMU<'static>>,
    pub RMT: Option<RMT<'static>>,
    pub RNG: Option<RNG<'static>>,
    pub RSA: Option<RSA<'static>>,
    pub SHA: Option<SHA<'static>>,
    pub SLCHOST: Option<SLCHOST<'static>>,
    pub ETM: Option<ETM<'static>>,
    pub SPI0: Option<SPI0<'static>>,
    pub SPI1: Option<SPI1<'static>>,
    pub SPI2: Option<SPI2<'static>>,
    pub SYSTEM: Option<SYSTEM<'static>>,
    pub SYSTIMER: Option<SYSTIMER<'static>>,
    pub TEE: Option<TEE<'static>>,
    pub TIMG0: Option<TIMG0<'static>>,
    pub TIMG1: Option<TIMG1<'static>>,
    pub TRACE0: Option<TRACE0<'static>>,
    pub TWAI0: Option<TWAI0<'static>>,
    pub TWAI1: Option<TWAI1<'static>>,
    pub UART0: Option<UART0<'static>>,
    pub UART1: Option<UART1<'static>>,
    pub UHCI0: Option<UHCI0<'static>>,
    pub USB_DEVICE: Option<USB_DEVICE<'static>>,
    pub DMA_CH0: Option<DMA_CH0<'static>>,
    pub DMA_CH1: Option<DMA_CH1<'static>>,
    pub DMA_CH2: Option<DMA_CH2<'static>>,
    pub ADC1: Option<ADC1<'static>>,
    pub BT: Option<BT<'static>>,
    pub FLASH: Option<FLASH<'static>>,
    pub LP_CORE: Option<LP_CORE<'static>>,
    pub SW_INTERRUPT: Option<SW_INTERRUPT<'static>>,
    pub TSENS: Option<TSENS<'static>>,
    pub WIFI: Option<WIFI<'static>>,
    pub MEM2MEM1: Option<MEM2MEM1<'static>>,
    pub MEM2MEM4: Option<MEM2MEM4<'static>>,
    pub MEM2MEM5: Option<MEM2MEM5<'static>>,
    pub MEM2MEM10: Option<MEM2MEM10<'static>>,
    pub MEM2MEM11: Option<MEM2MEM11<'static>>,
    pub MEM2MEM12: Option<MEM2MEM12<'static>>,
    pub MEM2MEM13: Option<MEM2MEM13<'static>>,
    pub MEM2MEM14: Option<MEM2MEM14<'static>>,
    pub MEM2MEM15: Option<MEM2MEM15<'static>>,

}

The Peripherals struct provides access to all of the hardware peripherals on the chip.

Fields
GPIO0: Option<GPIO0<'static>>

The GPIO0 peripheral.

GPIO1: Option<GPIO1<'static>>

The GPIO1 peripheral.

GPIO2: Option<GPIO2<'static>>

The GPIO2 peripheral.

GPIO3: Option<GPIO3<'static>>

The GPIO3 peripheral.

GPIO4: Option<GPIO4<'static>>

The GPIO4 peripheral.

GPIO5: Option<GPIO5<'static>>

The GPIO5 peripheral.

GPIO6: Option<GPIO6<'static>>

The GPIO6 peripheral.

GPIO7: Option<GPIO7<'static>>

The GPIO7 peripheral.

GPIO8: Option<GPIO8<'static>>

The GPIO8 peripheral.

GPIO9: Option<GPIO9<'static>>

The GPIO9 peripheral.

GPIO10: Option<GPIO10<'static>>

The GPIO10 peripheral.

GPIO11: Option<GPIO11<'static>>

The GPIO11 peripheral.

GPIO12: Option<GPIO12<'static>>

The GPIO12 peripheral.

GPIO13: Option<GPIO13<'static>>

The GPIO13 peripheral.

GPIO14: Option<GPIO14<'static>>

The GPIO14 peripheral.

GPIO15: Option<GPIO15<'static>>

The GPIO15 peripheral.

GPIO16: Option<GPIO16<'static>>

The GPIO16 peripheral.

GPIO17: Option<GPIO17<'static>>

The GPIO17 peripheral.

GPIO18: Option<GPIO18<'static>>

The GPIO18 peripheral.

GPIO19: Option<GPIO19<'static>>

The GPIO19 peripheral.

GPIO20: Option<GPIO20<'static>>

The GPIO20 peripheral.

GPIO21: Option<GPIO21<'static>>

The GPIO21 peripheral.

GPIO22: Option<GPIO22<'static>>

The GPIO22 peripheral.

GPIO23: Option<GPIO23<'static>>

The GPIO23 peripheral.

GPIO27: Option<GPIO27<'static>>

The GPIO27 peripheral.

AES: Option<AES<'static>>
Available on crate feature unstable only.

The AES peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

APB_SARADC: Option<APB_SARADC<'static>>
Available on crate feature unstable only.

The APB_SARADC peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

ASSIST_DEBUG: Option<ASSIST_DEBUG<'static>>
Available on crate feature unstable only.

The ASSIST_DEBUG peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

ATOMIC: Option<ATOMIC<'static>>
Available on crate feature unstable only.

The ATOMIC peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

DMA: Option<DMA<'static>>
Available on crate feature unstable only.

The DMA peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

DS: Option<DS<'static>>
Available on crate feature unstable only.

The DS peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

ECC: Option<ECC<'static>>
Available on crate feature unstable only.

The ECC peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

EFUSE: Option<EFUSE<'static>>
Available on crate feature unstable only.

The EFUSE peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

EXTMEM: Option<EXTMEM<'static>>
Available on crate feature unstable only.

The EXTMEM peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

GPIO: Option<GPIO<'static>>
Available on crate feature unstable only.

The GPIO peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

GPIO_SD: Option<GPIO_SD<'static>>
Available on crate feature unstable only.

The GPIO_SD peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

HINF: Option<HINF<'static>>
Available on crate feature unstable only.

The HINF peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

HMAC: Option<HMAC<'static>>
Available on crate feature unstable only.

The HMAC peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

HP_APM: Option<HP_APM<'static>>
Available on crate feature unstable only.

The HP_APM peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

HP_SYS: Option<HP_SYS<'static>>
Available on crate feature unstable only.

The HP_SYS peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

I2C_ANA_MST: Option<I2C_ANA_MST<'static>>
Available on crate feature unstable only.

The I2C_ANA_MST peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

I2C0: Option<I2C0<'static>>

The I2C0 peripheral.

I2S0: Option<I2S0<'static>>
Available on crate feature unstable only.

The I2S0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

IEEE802154: Option<IEEE802154<'static>>
Available on crate feature unstable only.

The IEEE802154 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

INTERRUPT_CORE0: Option<INTERRUPT_CORE0<'static>>
Available on crate feature unstable only.

The INTERRUPT_CORE0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

INTPRI: Option<INTPRI<'static>>
Available on crate feature unstable only.

The INTPRI peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

IO_MUX: Option<IO_MUX<'static>>
Available on crate feature unstable only.

The IO_MUX peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LEDC: Option<LEDC<'static>>
Available on crate feature unstable only.

The LEDC peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_ANA: Option<LP_ANA<'static>>
Available on crate feature unstable only.

The LP_ANA peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_AON: Option<LP_AON<'static>>
Available on crate feature unstable only.

The LP_AON peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_APM: Option<LP_APM<'static>>
Available on crate feature unstable only.

The LP_APM peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_APM0: Option<LP_APM0<'static>>
Available on crate feature unstable only.

The LP_APM0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_CLKRST: Option<LP_CLKRST<'static>>
Available on crate feature unstable only.

The LP_CLKRST peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_I2C0: Option<LP_I2C0<'static>>
Available on crate feature unstable only.

The LP_I2C0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_I2C_ANA_MST: Option<LP_I2C_ANA_MST<'static>>
Available on crate feature unstable only.

The LP_I2C_ANA_MST peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_IO: Option<LP_IO<'static>>
Available on crate feature unstable only.

The LP_IO peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_PERI: Option<LP_PERI<'static>>
Available on crate feature unstable only.

The LP_PERI peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_TEE: Option<LP_TEE<'static>>
Available on crate feature unstable only.

The LP_TEE peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_TIMER: Option<LP_TIMER<'static>>
Available on crate feature unstable only.

The LP_TIMER peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_UART: Option<LP_UART<'static>>
Available on crate feature unstable only.

The LP_UART peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_WDT: Option<LP_WDT<'static>>
Available on crate feature unstable only.

The LP_WDT peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LPWR: Option<LPWR<'static>>
Available on crate feature unstable only.

The LPWR peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MCPWM0: Option<MCPWM0<'static>>
Available on crate feature unstable only.

The MCPWM0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM_MONITOR: Option<MEM_MONITOR<'static>>
Available on crate feature unstable only.

The MEM_MONITOR peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MODEM_LPCON: Option<MODEM_LPCON<'static>>
Available on crate feature unstable only.

The MODEM_LPCON peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MODEM_SYSCON: Option<MODEM_SYSCON<'static>>
Available on crate feature unstable only.

The MODEM_SYSCON peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

OTP_DEBUG: Option<OTP_DEBUG<'static>>
Available on crate feature unstable only.

The OTP_DEBUG peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PARL_IO: Option<PARL_IO<'static>>
Available on crate feature unstable only.

The PARL_IO peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PAU: Option<PAU<'static>>
Available on crate feature unstable only.

The PAU peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PCNT: Option<PCNT<'static>>
Available on crate feature unstable only.

The PCNT peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PCR: Option<PCR<'static>>
Available on crate feature unstable only.

The PCR peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PLIC_MX: Option<PLIC_MX<'static>>
Available on crate feature unstable only.

The PLIC_MX peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

PMU: Option<PMU<'static>>
Available on crate feature unstable only.

The PMU peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

RMT: Option<RMT<'static>>
Available on crate feature unstable only.

The RMT peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

RNG: Option<RNG<'static>>
Available on crate feature unstable only.

The RNG peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

RSA: Option<RSA<'static>>
Available on crate feature unstable only.

The RSA peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SHA: Option<SHA<'static>>
Available on crate feature unstable only.

The SHA peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SLCHOST: Option<SLCHOST<'static>>
Available on crate feature unstable only.

The SLCHOST peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

ETM: Option<ETM<'static>>
Available on crate feature unstable only.

The ETM peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SPI0: Option<SPI0<'static>>
Available on crate feature unstable only.

The SPI0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SPI1: Option<SPI1<'static>>
Available on crate feature unstable only.

The SPI1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SPI2: Option<SPI2<'static>>

The SPI2 peripheral.

SYSTEM: Option<SYSTEM<'static>>
Available on crate feature unstable only.

The SYSTEM peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SYSTIMER: Option<SYSTIMER<'static>>
Available on crate feature unstable only.

The SYSTIMER peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TEE: Option<TEE<'static>>
Available on crate feature unstable only.

The TEE peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TIMG0: Option<TIMG0<'static>>
Available on crate feature unstable only.

The TIMG0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TIMG1: Option<TIMG1<'static>>
Available on crate feature unstable only.

The TIMG1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TRACE0: Option<TRACE0<'static>>
Available on crate feature unstable only.

The TRACE0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TWAI0: Option<TWAI0<'static>>
Available on crate feature unstable only.

The TWAI0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TWAI1: Option<TWAI1<'static>>
Available on crate feature unstable only.

The TWAI1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

UART0: Option<UART0<'static>>

The UART0 peripheral.

UART1: Option<UART1<'static>>

The UART1 peripheral.

UHCI0: Option<UHCI0<'static>>
Available on crate feature unstable only.

The UHCI0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

USB_DEVICE: Option<USB_DEVICE<'static>>
Available on crate feature unstable only.

The USB_DEVICE peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

DMA_CH0: Option<DMA_CH0<'static>>
Available on crate feature unstable only.

The DMA_CH0 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

DMA_CH1: Option<DMA_CH1<'static>>
Available on crate feature unstable only.

The DMA_CH1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

DMA_CH2: Option<DMA_CH2<'static>>
Available on crate feature unstable only.

The DMA_CH2 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

ADC1: Option<ADC1<'static>>
Available on crate feature unstable only.

The ADC1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

BT: Option<BT<'static>>
Available on crate feature unstable only.

The BT peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

FLASH: Option<FLASH<'static>>
Available on crate feature unstable only.

The FLASH peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

LP_CORE: Option<LP_CORE<'static>>
Available on crate feature unstable only.

The LP_CORE peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

SW_INTERRUPT: Option<SW_INTERRUPT<'static>>
Available on crate feature unstable only.

The SW_INTERRUPT peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

TSENS: Option<TSENS<'static>>
Available on crate feature unstable only.

The TSENS peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

WIFI: Option<WIFI<'static>>
Available on crate feature unstable only.

The WIFI peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM1: Option<MEM2MEM1<'static>>
Available on crate feature unstable only.

The MEM2MEM1 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM4: Option<MEM2MEM4<'static>>
Available on crate feature unstable only.

The MEM2MEM4 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM5: Option<MEM2MEM5<'static>>
Available on crate feature unstable only.

The MEM2MEM5 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM10: Option<MEM2MEM10<'static>>
Available on crate feature unstable only.

The MEM2MEM10 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM11: Option<MEM2MEM11<'static>>
Available on crate feature unstable only.

The MEM2MEM11 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM12: Option<MEM2MEM12<'static>>
Available on crate feature unstable only.

The MEM2MEM12 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM13: Option<MEM2MEM13<'static>>
Available on crate feature unstable only.

The MEM2MEM13 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM14: Option<MEM2MEM14<'static>>
Available on crate feature unstable only.

The MEM2MEM14 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.

MEM2MEM15: Option<MEM2MEM15<'static>>
Available on crate feature unstable only.

The MEM2MEM15 peripheral. This API is marked as unstable and is only available when the unstable crate feature is enabled. This comes with no stability guarantees, and could be changed or removed at any time.