ariel_os_esp::peripherals

Struct Peripherals

pub struct Peripherals {

    pub ADC1: ADC1,
    pub AES: AES,
    pub ASSIST_DEBUG: ASSIST_DEBUG,
    pub ATOMIC: ATOMIC,
    pub BT: BT,
    pub DS: DS,
    pub ECC: ECC,
    pub EFUSE: EFUSE,
    pub EXTMEM: EXTMEM,
    pub GPIO_SD: GPIO_SD,
    pub HINF: HINF,
    pub HMAC: HMAC,
    pub HP_APM: HP_APM,
    pub HP_SYS: HP_SYS,
    pub I2C0: I2C0,
    pub I2S0: I2S0,
    pub IEEE802154: IEEE802154,
    pub INTERRUPT_CORE0: INTERRUPT_CORE0,
    pub INTPRI: INTPRI,
    pub IO_MUX: IO_MUX,
    pub LEDC: LEDC,
    pub LPWR: LPWR,
    pub LP_CORE: LP_CORE,
    pub LP_PERI: LP_PERI,
    pub LP_ANA: LP_ANA,
    pub LP_AON: LP_AON,
    pub LP_APM: LP_APM,
    pub LP_APM0: LP_APM0,
    pub LP_I2C0: LP_I2C0,
    pub LP_I2C_ANA_MST: LP_I2C_ANA_MST,
    pub LP_IO: LP_IO,
    pub LP_TEE: LP_TEE,
    pub LP_TIMER: LP_TIMER,
    pub LP_UART: LP_UART,
    pub LP_WDT: LP_WDT,
    pub MCPWM0: MCPWM0,
    pub MEM_MONITOR: MEM_MONITOR,
    pub OTP_DEBUG: OTP_DEBUG,
    pub PARL_IO: PARL_IO,
    pub PAU: PAU,
    pub PCNT: PCNT,
    pub PMU: PMU,
    pub RADIO_CLK: RADIO_CLK,
    pub RMT: RMT,
    pub RNG: RNG,
    pub RSA: RSA,
    pub SHA: SHA,
    pub SLCHOST: SLCHOST,
    pub SOC_ETM: SOC_ETM,
    pub SPI0: SPI0,
    pub SPI1: SPI1,
    pub SPI2: SPI2,
    pub SYSTEM: SYSTEM,
    pub SYSTIMER: SYSTIMER,
    pub SW_INTERRUPT: SW_INTERRUPT,
    pub TEE: TEE,
    pub TIMG0: TIMG0,
    pub TIMG1: TIMG1,
    pub TRACE0: TRACE0,
    pub TSENS: TSENS,
    pub TWAI0: TWAI0,
    pub TWAI1: TWAI1,
    pub UART0: UART0,
    pub UART1: UART1,
    pub UHCI0: UHCI0,
    pub USB_DEVICE: USB_DEVICE,
    pub WIFI: WIFI,
    pub MEM2MEM1: MEM2MEM1,
    pub MEM2MEM4: MEM2MEM4,
    pub MEM2MEM5: MEM2MEM5,
    pub MEM2MEM10: MEM2MEM10,
    pub MEM2MEM11: MEM2MEM11,
    pub MEM2MEM12: MEM2MEM12,
    pub MEM2MEM13: MEM2MEM13,
    pub MEM2MEM14: MEM2MEM14,
    pub MEM2MEM15: MEM2MEM15,
    pub GPIO0: GpioPin<0>,
    pub GPIO1: GpioPin<1>,
    pub GPIO2: GpioPin<2>,
    pub GPIO3: GpioPin<3>,
    pub GPIO4: GpioPin<4>,
    pub GPIO5: GpioPin<5>,
    pub GPIO6: GpioPin<6>,
    pub GPIO7: GpioPin<7>,
    pub GPIO8: GpioPin<8>,
    pub GPIO9: GpioPin<9>,
    pub GPIO10: GpioPin<10>,
    pub GPIO11: GpioPin<11>,
    pub GPIO12: GpioPin<12>,
    pub GPIO13: GpioPin<13>,
    pub GPIO14: GpioPin<14>,
    pub GPIO15: GpioPin<15>,
    pub GPIO16: GpioPin<16>,
    pub GPIO17: GpioPin<17>,
    pub GPIO18: GpioPin<18>,
    pub GPIO19: GpioPin<19>,
    pub GPIO20: GpioPin<20>,
    pub GPIO21: GpioPin<21>,
    pub GPIO22: GpioPin<22>,
    pub GPIO23: GpioPin<23>,
    pub GPIO24: GpioPin<24>,
    pub GPIO25: GpioPin<25>,
    pub GPIO26: GpioPin<26>,
    pub GPIO27: GpioPin<27>,
    pub GPIO28: GpioPin<28>,
    pub GPIO29: GpioPin<29>,
    pub GPIO30: GpioPin<30>,
    pub DMA_CH0: DmaChannel0,
    pub DMA_CH1: DmaChannel1,
    pub DMA_CH2: DmaChannel2,

}

The Peripherals struct provides access to all of the hardware peripherals on the chip.

Fields
ADC1: ADC1

The ADC1 peripheral.

AES: AES

The AES peripheral.

ASSIST_DEBUG: ASSIST_DEBUG

The ASSIST_DEBUG peripheral.

ATOMIC: ATOMIC

The ATOMIC peripheral.

BT: BT

The BT peripheral.

DS: DS

The DS peripheral.

ECC: ECC

The ECC peripheral.

EFUSE: EFUSE

The EFUSE peripheral.

EXTMEM: EXTMEM

The EXTMEM peripheral.

GPIO_SD: GPIO_SD

The GPIO_SD peripheral.

HINF: HINF

The HINF peripheral.

HMAC: HMAC

The HMAC peripheral.

HP_APM: HP_APM

The HP_APM peripheral.

HP_SYS: HP_SYS

The HP_SYS peripheral.

I2C0: I2C0

The I2C0 peripheral.

I2S0: I2S0

The I2S0 peripheral.

IEEE802154: IEEE802154

The IEEE802154 peripheral.

INTERRUPT_CORE0: INTERRUPT_CORE0

The INTERRUPT_CORE0 peripheral.

INTPRI: INTPRI

The INTPRI peripheral.

IO_MUX: IO_MUX

The IO_MUX peripheral.

LEDC: LEDC

The LEDC peripheral.

LPWR: LPWR

The LPWR peripheral.

LP_CORE: LP_CORE

The LP_CORE peripheral.

LP_PERI: LP_PERI

The LP_PERI peripheral.

LP_ANA: LP_ANA

The LP_ANA peripheral.

LP_AON: LP_AON

The LP_AON peripheral.

LP_APM: LP_APM

The LP_APM peripheral.

LP_APM0: LP_APM0

The LP_APM0 peripheral.

LP_I2C0: LP_I2C0

The LP_I2C0 peripheral.

LP_I2C_ANA_MST: LP_I2C_ANA_MST

The LP_I2C_ANA_MST peripheral.

LP_IO: LP_IO

The LP_IO peripheral.

LP_TEE: LP_TEE

The LP_TEE peripheral.

LP_TIMER: LP_TIMER

The LP_TIMER peripheral.

LP_UART: LP_UART

The LP_UART peripheral.

LP_WDT: LP_WDT

The LP_WDT peripheral.

MCPWM0: MCPWM0

The MCPWM0 peripheral.

MEM_MONITOR: MEM_MONITOR

The MEM_MONITOR peripheral.

OTP_DEBUG: OTP_DEBUG

The OTP_DEBUG peripheral.

PARL_IO: PARL_IO

The PARL_IO peripheral.

PAU: PAU

The PAU peripheral.

PCNT: PCNT

The PCNT peripheral.

PMU: PMU

The PMU peripheral.

RADIO_CLK: RADIO_CLK

The RADIO_CLK peripheral.

RMT: RMT

The RMT peripheral.

RNG: RNG

The RNG peripheral.

RSA: RSA

The RSA peripheral.

SHA: SHA

The SHA peripheral.

SLCHOST: SLCHOST

The SLCHOST peripheral.

SOC_ETM: SOC_ETM

The SOC_ETM peripheral.

SPI0: SPI0

The SPI0 peripheral.

SPI1: SPI1

The SPI1 peripheral.

SPI2: SPI2

The SPI2 peripheral.

SYSTEM: SYSTEM

The SYSTEM peripheral.

SYSTIMER: SYSTIMER

The SYSTIMER peripheral.

SW_INTERRUPT: SW_INTERRUPT

The SW_INTERRUPT peripheral.

TEE: TEE

The TEE peripheral.

TIMG0: TIMG0

The TIMG0 peripheral.

TIMG1: TIMG1

The TIMG1 peripheral.

TRACE0: TRACE0

The TRACE0 peripheral.

TSENS: TSENS

The TSENS peripheral.

TWAI0: TWAI0

The TWAI0 peripheral.

TWAI1: TWAI1

The TWAI1 peripheral.

UART0: UART0

The UART0 peripheral.

UART1: UART1

The UART1 peripheral.

UHCI0: UHCI0

The UHCI0 peripheral.

USB_DEVICE: USB_DEVICE

The USB_DEVICE peripheral.

WIFI: WIFI

The WIFI peripheral.

MEM2MEM1: MEM2MEM1

The MEM2MEM1 peripheral.

MEM2MEM4: MEM2MEM4

The MEM2MEM4 peripheral.

MEM2MEM5: MEM2MEM5

The MEM2MEM5 peripheral.

MEM2MEM10: MEM2MEM10

The MEM2MEM10 peripheral.

MEM2MEM11: MEM2MEM11

The MEM2MEM11 peripheral.

MEM2MEM12: MEM2MEM12

The MEM2MEM12 peripheral.

MEM2MEM13: MEM2MEM13

The MEM2MEM13 peripheral.

MEM2MEM14: MEM2MEM14

The MEM2MEM14 peripheral.

MEM2MEM15: MEM2MEM15

The MEM2MEM15 peripheral.

GPIO0: GpioPin<0>

GPIO0

GPIO1: GpioPin<1>

GPIO1

GPIO2: GpioPin<2>

GPIO2

GPIO3: GpioPin<3>

GPIO3

GPIO4: GpioPin<4>

GPIO4

GPIO5: GpioPin<5>

GPIO5

GPIO6: GpioPin<6>

GPIO6

GPIO7: GpioPin<7>

GPIO7

GPIO8: GpioPin<8>

GPIO8

GPIO9: GpioPin<9>

GPIO9

GPIO10: GpioPin<10>

GPIO10

GPIO11: GpioPin<11>

GPIO11

GPIO12: GpioPin<12>

GPIO12

GPIO13: GpioPin<13>

GPIO13

GPIO14: GpioPin<14>

GPIO14

GPIO15: GpioPin<15>

GPIO15

GPIO16: GpioPin<16>

GPIO16

GPIO17: GpioPin<17>

GPIO17

GPIO18: GpioPin<18>

GPIO18

GPIO19: GpioPin<19>

GPIO19

GPIO20: GpioPin<20>

GPIO20

GPIO21: GpioPin<21>

GPIO21

GPIO22: GpioPin<22>

GPIO22

GPIO23: GpioPin<23>

GPIO23

GPIO24: GpioPin<24>

GPIO24

GPIO25: GpioPin<25>

GPIO25

GPIO26: GpioPin<26>

GPIO26

GPIO27: GpioPin<27>

GPIO27

GPIO28: GpioPin<28>

GPIO28

GPIO29: GpioPin<29>

GPIO29

GPIO30: GpioPin<30>

GPIO30

DMA_CH0: DmaChannel0

DMA_CH0 DMA channel.

DMA_CH1: DmaChannel1

DMA_CH1 DMA channel.

DMA_CH2: DmaChannel2

DMA_CH2 DMA channel.

Implementations

impl Peripherals

pub unsafe fn steal() -> Peripherals

Unsafely create an instance of this peripheral out of thin air.

Safety

You must ensure that you’re only using one instance of this type at a time.