ariel_os_esp

Module peripherals

Types for the peripheral singletons.

Structs

ADC1

ADC1 peripheral singleton

AES

AES peripheral singleton

ASSIST_DEBUG

ASSIST_DEBUG peripheral singleton

ATOMIC

ATOMIC peripheral singleton

BT

BT peripheral singleton

DS

DS peripheral singleton

ECC

ECC peripheral singleton

EFUSE

EFUSE peripheral singleton

EXTMEM

EXTMEM peripheral singleton

GPIO_SD

GPIO_SD peripheral singleton

HINF

HINF peripheral singleton

HMAC

HMAC peripheral singleton

HP_APM

HP_APM peripheral singleton

HP_SYS

HP_SYS peripheral singleton

I2C0

I2C0 peripheral singleton

I2S0

I2S0 peripheral singleton

IEEE802154

IEEE802154 peripheral singleton

INTERRUPT_CORE0

INTERRUPT_CORE0 peripheral singleton

INTPRI

INTPRI peripheral singleton

IO_MUX

IO_MUX peripheral singleton

LEDC

LEDC peripheral singleton

LPWR

LPWR peripheral singleton

LP_ANA

LP_ANA peripheral singleton

LP_AON

LP_AON peripheral singleton

LP_APM

LP_APM peripheral singleton

LP_APM0

LP_APM0 peripheral singleton

LP_CORE

LP_CORE peripheral singleton

LP_I2C0

LP_I2C0 peripheral singleton

LP_I2C_ANA_MST

LP_I2C_ANA_MST peripheral singleton

LP_IO

LP_IO peripheral singleton

LP_PERI

LP_PERI peripheral singleton

LP_TEE

LP_TEE peripheral singleton

LP_TIMER

LP_TIMER peripheral singleton

LP_UART

LP_UART peripheral singleton

LP_WDT

LP_WDT peripheral singleton

MCPWM0

MCPWM0 peripheral singleton

MEM2MEM1

MEM2MEM1 peripheral singleton

MEM2MEM4

MEM2MEM4 peripheral singleton

MEM2MEM5

MEM2MEM5 peripheral singleton

MEM2MEM10

MEM2MEM10 peripheral singleton

MEM2MEM11

MEM2MEM11 peripheral singleton

MEM2MEM12

MEM2MEM12 peripheral singleton

MEM2MEM13

MEM2MEM13 peripheral singleton

MEM2MEM14

MEM2MEM14 peripheral singleton

MEM2MEM15

MEM2MEM15 peripheral singleton

MEM_MONITOR

MEM_MONITOR peripheral singleton

OTP_DEBUG

OTP_DEBUG peripheral singleton

OptionalPeripherals

The Peripherals struct provides access to all of the hardware peripherals on the chip.

PARL_IO

PARL_IO peripheral singleton

PAU

PAU peripheral singleton

PCNT

PCNT peripheral singleton

PMU

PMU peripheral singleton

Peripherals

The Peripherals struct provides access to all of the hardware peripherals on the chip.

RADIO_CLK

RADIO_CLK peripheral singleton

RMT

RMT peripheral singleton

RNG

RNG peripheral singleton

RSA

RSA peripheral singleton

SHA

SHA peripheral singleton

SLCHOST

SLCHOST peripheral singleton

SOC_ETM

SOC_ETM peripheral singleton

SPI0

SPI0 peripheral singleton

SPI1

SPI1 peripheral singleton

SPI2

SPI2 peripheral singleton

SW_INTERRUPT

SW_INTERRUPT peripheral singleton

SYSTEM

SYSTEM peripheral singleton

SYSTIMER

SYSTIMER peripheral singleton

TEE

TEE peripheral singleton

TIMG0

TIMG0 peripheral singleton

TIMG1

TIMG1 peripheral singleton

TRACE0

TRACE0 peripheral singleton

TSENS

TSENS peripheral singleton

TWAI0

TWAI0 peripheral singleton

TWAI1

TWAI1 peripheral singleton

UART0

UART0 peripheral singleton

UART1

UART1 peripheral singleton

UHCI0

UHCI0 peripheral singleton

USB_DEVICE

USB_DEVICE peripheral singleton

WIFI

WIFI peripheral singleton

Enums

Interrupt

Enumeration of all the interrupts.

Type Aliases

GPIO0

GPIO1

GPIO2

GPIO3

GPIO4

GPIO5

GPIO6

GPIO7

GPIO8

GPIO9

GPIO10

GPIO11

GPIO12

GPIO13

GPIO14

GPIO15

GPIO16

GPIO17

GPIO18

GPIO19

GPIO20

GPIO21

GPIO22

GPIO23

GPIO24

GPIO25

GPIO26

GPIO27

GPIO28

GPIO29

GPIO30